extract interface into api
