+++ lib/freshbooks_api_client/interface.ex
@@ -225,56 +225,4 @@ defmodule FreshbooksApiClient.Interface do
   def to_schema(schema, params) do
     struct!(schema, params)
   end
-
-  def call(interface, method, params \\ []) do
-    use Retry
-
-    retry with: [5_000, 30_000, 60_000], rescue_only: [FreshbooksApiClient.RateLimitError] do
-      call_without_retry(interface, method, params)
-    end
-  end
-
-  # TODO: I'm not sure what a rate limit error looks like, but when we get one we need to raise this exception
-  def call_without_retry(interface, method, params) do
-    apply(interface, method, [FreshbooksApiClient.caller(), params])
-  end
-
-  def all(interface, params \\ []) do
-    use Retry
-
-    retry with: [5_000, 30_000, 60_000], rescue_only: [FreshbooksApiClient.PaginationError] do
-      all_without_retry(interface, params)
-    end
-  end
-
-  # Failure Scenarios (per_page: 3)
-  # 1,2,3 + 4,5,6 total == 6
-
-  # delete 3 after page 1
-  # 1,2,3, + 5,6 total == 5 (total count changed)
-
-  # add 7 after page 1
-  # 1,2,3 + 4,5,6 + 7 total == 7 (total count & total_pages changed - we could just grab another page)
-
-  # delete 3, add 7 after page 1
-  # 1,2,3 + 5,6,7 total == 6 (the results are inaccurate - Check that #4 is actually deleted w/ a get call to the API)
-  def all_without_retry(interface, params) do
-    results = call(interface, :list, [Keyword.merge(params, [per_page: 100])])
-
-    pages = results[:pages]
-    total = results[:total]
-    resources = results[:resources]
-
-    if pages == 1 do
-      resources
-    else
-      Enum.reduce(Range.new(2, pages), resources, fn(page, acc) ->
-        results = call(interface, :list, [Keyword.merge(params, [per_page: 100, page: page])])
-
-        if results[:total] != total, do: raise FreshbooksApiClient.PaginationError
-
-        acc ++ results[:resources]
-      end)
-    end
-  end
 end
